UBUNTU: [Config] Set the default CPU governor to ONDEMAND
BugLink: https://bugs.launchpad.net/bugs/1885730

The cpu-freq script in systemd has been removed for Ubuntu 20.10
because of conflicts between users wanting to change cpu governor
state and the defaulted post-boot ondemand setting my be set at
the inappropriate time post-boot (or with a setting that the user
doesn't want).  To remove all the confusing complexity, the best
viable option for now is to default to the ondemand governor and
let the user select their preferred setting post-boot.  As it
stands, the current performance governor is not power efficient
and not the best all round default, so it needs to be changed.

We can re-visit ways to speed up kernel boot times with governor
settings at a later date.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>